backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.0
azure-storage-blob==12.22.0
Django==4.2.23
django-ninja==1.4.3
pydantic==2.11.7
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
openai>=1.0.0
instructor>=0.6.0
requests>=2.28.0
pandas>=1.5.0
